requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.5
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-nine==0.2.4
djangorestframework==3.12.2
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
